✨ TRACKER-9: Use BigDecimal for kg and long for g
This avoids floating point errors. The database stores weight in grams,
the domain represents the weight in grams with 50g increments, and
the application uses the weight in kilograms.

The types:
- INTEGER in database
- BigDecimal in application
- long in domain

Flyway migration scripts were edited, with the same idea as in
2eb5ce13ff88ef81b85ae85984a91f2c26979ce3 and
a1b1509c9dfc61d558b09a18952a804f7f81a303 .

Closes #1.

+++ backend/src/main/java/fr/sylvainjanet/tracker/journal/adapter/in/web/dtos/request/LogWeightMeasurementRequest.java
@@ -2,7 +2,8 @@ package fr.sylvainjanet.tracker.journal.adapter.in.web.dtos.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record LogWeightMeasurementRequest(
-        @NotNull LocalDate date, @NotNull @Positive Float weightInKg) {}
+        @NotNull LocalDate date, @NotNull @Positive BigDecimal weightInKg) {}
